Add basic fields to data model classes
The intent of the BusInfo class is the data about what the user is
interested in. This encompasses the name of the info, the color, as well
as the IDs of content in the OCTranspo API (stop, bus, and direction).

The Arrivals object is intended to store data about when buses will
arrive and how old the data is.